move precompiles to celo_contracts.go
