+++ core/vm/contracts_test.go
@@ -186,37 +186,37 @@ var allPrecompiles = map[common.Address]PrecompiledContract{
 	common.BytesToAddress([]byte{18}):         &bls12381MapG2{},
 
 	// Celo Precompiled Contracts
-	transferAddress:              &transfer{},
-	fractionMulExpAddress:        &fractionMulExp{},
-	proofOfPossessionAddress:     &proofOfPossession{},
-	getValidatorAddress:          &getValidator{},
-	numberValidatorsAddress:      &numberValidators{},
-	epochSizeAddress:             &epochSize{},
-	blockNumberFromHeaderAddress: &blockNumberFromHeader{},
-	hashHeaderAddress:            &hashHeader{},
-	getParentSealBitmapAddress:   &getParentSealBitmap{},
-	getVerifiedSealBitmapAddress: &getVerifiedSealBitmap{},
+	celoPrecompileAddress(2):  &transfer{},
+	celoPrecompileAddress(3):  &fractionMulExp{},
+	celoPrecompileAddress(4):  &proofOfPossession{},
+	celoPrecompileAddress(5):  &getValidator{},
+	celoPrecompileAddress(6):  &numberValidators{},
+	celoPrecompileAddress(7):  &epochSize{},
+	celoPrecompileAddress(8):  &blockNumberFromHeader{},
+	celoPrecompileAddress(9):  &hashHeader{},
+	celoPrecompileAddress(10): &getParentSealBitmap{},
+	celoPrecompileAddress(11): &getVerifiedSealBitmap{},
 
 	// New in Donut hard fork
-	ed25519Address:           &ed25519Verify{},
-	b12_381G1AddAddress:      &bls12381G1Add{},
-	b12_381G1MulAddress:      &bls12381G1Mul{},
-	b12_381G1MultiExpAddress: &bls12381G1MultiExp{},
-	b12_381G2AddAddress:      &bls12381G2Add{},
-	b12_381G2MulAddress:      &bls12381G2Mul{},
-	b12_381G2MultiExpAddress: &bls12381G2MultiExp{},
-	b12_381PairingAddress:    &bls12381Pairing{},
-	b12_381MapFpToG1Address:  &bls12381MapG1{},
-	b12_381MapFp2ToG2Address: &bls12381MapG2{},
-	b12_377G1AddAddress:      &bls12377G1Add{},
-	b12_377G1MulAddress:      &bls12377G1Mul{},
-	b12_377G1MultiExpAddress: &bls12377G1MultiExp{},
-	b12_377G2AddAddress:      &bls12377G2Add{},
-	b12_377G2MulAddress:      &bls12377G2Mul{},
-	b12_377G2MultiExpAddress: &bls12377G2MultiExp{},
-	b12_377PairingAddress:    &bls12377Pairing{},
-	cip20Address:             &cip20HashFunctions{Cip20HashesDonut},
-	cip26Address:             &getValidatorBLS{},
+	celoPrecompileAddress(12): &ed25519Verify{},
+	celoPrecompileAddress(13): &bls12381G1Add{},
+	celoPrecompileAddress(14): &bls12381G1Mul{},
+	celoPrecompileAddress(15): &bls12381G1MultiExp{},
+	celoPrecompileAddress(16): &bls12381G2Add{},
+	celoPrecompileAddress(17): &bls12381G2Mul{},
+	celoPrecompileAddress(18): &bls12381G2MultiExp{},
+	celoPrecompileAddress(19): &bls12381Pairing{},
+	celoPrecompileAddress(20): &bls12381MapG1{},
+	celoPrecompileAddress(21): &bls12381MapG2{},
+	celoPrecompileAddress(22): &bls12377G1Add{},
+	celoPrecompileAddress(23): &bls12377G1Mul{},
+	celoPrecompileAddress(24): &bls12377G1MultiExp{},
+	celoPrecompileAddress(25): &bls12377G2Add{},
+	celoPrecompileAddress(26): &bls12377G2Mul{},
+	celoPrecompileAddress(27): &bls12377G2MultiExp{},
+	celoPrecompileAddress(28): &bls12377Pairing{},
+	celoPrecompileAddress(29): &cip20HashFunctions{Cip20HashesDonut},
+	celoPrecompileAddress(30): &getValidatorBLS{},
 }
 
 func testJSON(name, addr string, t *testing.T) {
